requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
numpy==1.24.5
Pillow==10.0.1
opencv-python==4.8.1.78
scikit-image==0.21.0
